一部カラムをbigintに型変更

+++ myapp/db/migrate/20230203061017_create_tasks.rb
@@ -1,7 +1,7 @@
 class CreateTasks < ActiveRecord::Migration[7.0]
   def change
     create_table :tasks do |t|
-      t.integer :owner_id, foreign_key: true, :null => false
+      t.bigint :owner_id, foreign_key: true, :null => false
       t.string :status, :null => false, comment: '["waiting", "doing", "completed"]'
       t.string :title, :null => false
       t.integer :priority, limit: 1, :null => false, comment: '{1: "high", 2: "middle", 3: "low"}'
